% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{add_fixrates}
\alias{add_fixrates}
\title{Add Fix Rates to Collar-Viewer Input Data}
\usage{
add_fixrates(dat, fxr_scheds)
}
\arguments{
\item{dat}{Input data.frame for Collar-Viewer}

\item{fxr_scheds}{Table of program schedules for GPS collar}
}
\value{
Input data.frame for Collar-Viewer including fix rate values
}
\description{
Add Fix Rates to Collar-Viewer Input Data
}
