% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{caps2animals}
\alias{caps2animals}
\title{Convert Lynx Capture Data to Animal Table for AniTrackTools}
\usage{
caps2animals(caps)
}
\arguments{
\item{caps}{A data.frame of lynx capture data imported from Access database(s)}
}
\value{
A data.frame formatted for use with AniTrackTools
}
\description{
Convert Lynx Capture Data to Animal Table for AniTrackTools
}
