% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tools.R
\name{find.octc}
\alias{find.octc}
\title{Helper Function to Find Collars Deployed on >1 Lynx}
\usage{
find.octc(captures)
}
\arguments{
\item{captures}{A data.frame object created from \code{import.captures} function.}
}
\value{
A data.frame containing row index, error message, and error code for capture records flagged with errors.
}
\description{
Helper Function to Find Collars Deployed on >1 Lynx
}
