% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_captures}
\alias{import_captures}
\title{Compile Capture Tables from Site-specific NWBF Lynx Access Databases}
\usage{
import_captures(
  file_names,
  telonics = TRUE,
  sites = c("kan", "kuk", "tet", "wsm", "ykf")
)
}
\arguments{
\item{file_names}{Character vector of full file paths to Access databases including file name}

\item{telonics}{Logical value indicating whether to only compile records for Telonics collars. Default is TRUE.}

\item{sites}{Character vector of site codes used to identify Access database file names to be used.}
}
\value{
A data.frame of compiled capture records.
}
\description{
Compile Capture Tables from Site-specific NWBF Lynx Access Databases
}
\examples{
\dontrun{
import_captures("tet_lynx_capture_database_202000101.accdb")}
}
