% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_table}
\alias{import_table}
\title{Import Tables from Access Database}
\usage{
import_table(full_file_path, table_name)
}
\arguments{
\item{full_file_path}{Directory path to Access database including file name}

\item{table_name}{Character vector of table names to be imported}
}
\value{
Data.frame
}
\description{
Import Tables from Access Database
}
