% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{make_cvdat}
\alias{make_cvdat}
\title{Create Collar-Viewer Input Data}
\usage{
make_cvdat(captures, locations)
}
\arguments{
\item{captures}{data.frame of capture data}

\item{locations}{gps data}
}
\value{
data.frame used as input for collar-viewer app
}
\description{
Create Collar-Viewer Input Data
}
