% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tools.R
\name{qc_captures}
\alias{qc_captures}
\title{Qualit Control Check Capture Data}
\usage{
qc_captures(captures)
}
\arguments{
\item{captures}{A data.frame object created from \code{import.captures} function}
}
\value{
A data.frame containing row index, error message, and error code for capture records flagged with errors.
}
\description{
Qualit Control Check Capture Data
}
