% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{refactor}
\alias{refactor}
\title{Recalculate Factor Levels Based on Unique Values}
\usage{
refactor(df)
}
\arguments{
\item{df}{Data.frame}
}
\value{
Data.frame
}
\description{
Recalculate Factor Levels Based on Unique Values
}
\examples{
df <-  data.frame(x = factor(LETTERS[1:5], levels = LETTERS[1:10]))
(length(unique(df$x))==nlevels(df$x))
df <-  refactor(df)
(length(unique(df$x))==nlevels(df$x))
}
