% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tools.R
\name{remove_capture_errors}
\alias{remove_capture_errors}
\title{Remove Lynx and Collars with Red-Light Errors}
\usage{
remove_capture_errors(captures, capture_errors)
}
\arguments{
\item{captures}{A data.frame object created from \code{import.captures} function.}

\item{capture_errors}{A data.frame object created from \code{qc_captures} function.}
}
\value{
A data.frame of capture records with "red-light" errors removed.
}
\description{
Remove Lynx and Collars with Red-Light Errors
}
\examples{
\dontrun{
captures <- import_captures(file_names)
capture_errors <- qc_captures(captures)
clean_captures <- remove_capture_errors(captures, capture_errors)}
}
